+++ frontend/src/createStream/SwipeableSteps/SelectDeviceSwipeableSteps.jsx
@@ -3,12 +3,12 @@ import PropTypes from "prop-types";
 
 import SwipeableViews from "react-swipeable-views";
 import { DialogTitle, MobileStepper } from "@material-ui/core";
-import SelectDeviceTable from "../SelectDeviceTable";
-import SelectChannelTable from "../SelectChannelTable";
-import StreamDeviceCard from "../../../general/StreamDeviceCard";
+import SelectDeviceTable from "../SelectDeviceDialog/SelectDeviceTable";
+import SelectChannelTable from "../SelectDeviceDialog/SelectChannelTable";
+import StreamDeviceCard from "../../general/StreamDeviceCard";
 
 import StepperNextButton from "./StepperNextButton";
-import DeviceInfo from "../../../model/DeviceInfo";
+import DeviceInfo from "../../model/DeviceInfo";
 import StepperBackButton from "./StepperBackButton";
 
 export default class SelectDeviceSwipeableSteps extends React.Component {
